services/payment-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-distro==0.50b0
